Task (http-client): rename utils to helpers for consistency
